Implement post to circle selector
— views/post/widgets/postForm.php
Shows a "post to space" select box when the criteria is met.

Criteria = if they're not looking at a Space or a User and they are a
member in more than one space.

Later on (in views/wall/widgets/contentForm.php) we set up a listener
on #post_to_space which updates the fields that tell Humhub where the
post is originating from.

— views/wall/widgets/contentForm.php
If they're looking at a Space or User, use default functionality. If
they're on the dashboard, implement "post to space" selector.

Once we've established they are on a "post to space" page and check if
the user is a member of more than one group. If the user is only in one
space, we pre-fill the fields and the select box is hidden.

If the user is in multiple spaces, we set up our listeners and disable
the post button until a selection is made

— views/user/profile/index.php
We no longer want to show the “post to profile” form.